refactor: move services and controllers to corresponding modules
