--- src/modules/user/create-user/create-user.ts
@@ -1,1 +1,0 @@
-export class CreateUser {}
