--- src/modules/user/user.controller.ts
@@ -1,15 +1,0 @@
-import { Body, Controller, Post } from '@nestjs/common';
-import { IsPublic } from 'src/modules/auth/decorators/is-public.decorator';
-import { CreateUserDto } from './dto/create-user.dto';
-import { UserService } from './user.service';
-
-@Controller('user')
-export class UserController {
-  constructor(private readonly userService: UserService) {}
-
-  @IsPublic()
-  @Post()
-  create(@Body() createUserDto: CreateUserDto) {
-    return this.userService.create(createUserDto);
-  }
-}
