+++ src/modules/user/user.module.ts
@@ -1,14 +1,13 @@
 import { Module } from '@nestjs/common';
 import { AppService } from 'src/app.service';
 import { PrismaModule } from 'src/prisma/prisma.module';
-import { UserController } from './user.controller';
-import { UserService } from './user.service';
-import { CreateUserService } from './create-user/create-user.service';
 import { CreateUserController } from './create-user/create-user.controller';
+import { CreateUserService } from './create-user/create-user.service';
+import { UserService } from './user.service';
 
 @Module({
   imports: [PrismaModule],
-  controllers: [UserController, CreateUserController],
+  controllers: [CreateUserController],
   providers: [UserService, AppService, CreateUserService],
   exports: [UserService],
 })
